WebHost: move imports around to save memory in MP
